src/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.12.3
groq==0.9.0
py-expression-eval==0.3.14
python-dotenv==1.0.1
requests==2.32.3
